website/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-bootstrap5==22.2
django-constance==3.1.0
django-grappelli==4.0.1
django-jazzmin==3.0.0
django-picklefield==3.2
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
